refactor(#155 PR 5/10): extract core/economy.js and core/hints.js from game.js
Pure code movement, zero behavior change. game.js: 3374 -> 2969 lines.

- src/core/economy.js (336 lines): updateRepairCostTable,
  updateFinancesDisplay, toggleAutoRepair, processAutoRepair,
  getAutoRepairUpkeep. window.toggleAutoRepair stays in game.js's
  boundary block, assigning the imported binding.
- src/core/hints.js (101 lines): checkSmartHints (exported) +
  showSmartHint (module-internal).

Deferred deliberately: the inline upkeep-charging block inside animate()
(not a named function — extracting it would mean authoring new code, not
moving it) and the thin window-assigned budget setters that share
game.js's syncInput helper. Both go with the loop/UI splits (PR 6+).

Verified: lint 0, 87/87 tests, 26-module graph links, in-browser —
smart hint fires from the new module, auto-repair toggles through the
window path with upkeep accruing, finances/repair tables render, 40/40
requests processed end-to-end, zero console errors.
